add a second it statement to filter list by tag tag
