+++ src/domUpdates.js
@@ -28,52 +28,46 @@ import {
   findRecipeByIngredient,
   findRecipeByName,
   getUserInput,
-} from "../src/scripts.js";
+  saveRecipe,
+  deleteRecipe,
+} from "../src/functions.js";
 
-import ingredientsData from "../src/data/ingredients.js";
+//any function nthat needs data, it gets passed the data when it's invoked
+/// change to functions
 
-import recipeData from "../src/data/recipes.js";
+// import ingredientsData from "../src/data/ingredients.js";
 
-import usersData from "../src/data/users.js";
+// import recipeData from "../src/data/recipes.js";
 
-const recipeDisplay = document.querySelector(".recipes");
+// import usersData from "../src/data/users.js";
 
-const modal = document.querySelector(".modal");
-const modalContainer = document.querySelector(".modal-container");
-const modalOverlay = document.querySelector(".modal-overlay");
-const modalTitle = document.querySelector(".modal-title");
-const modalTags = document.querySelector(".modal-tags");
-const modalDirections = document.querySelector(".modal-directions-list");
-const modalCost = document.querySelector(".modal-cost");
-const modalIngredients = document.querySelector(".modal-ingredients-list");
-const closeBtn = document.querySelector(".close-btn");
-const tagButtons = document.querySelector(".tag-buttons");
-const inputName = document.querySelector(".input-name");
-const inputIngredient = document.querySelector(".input-ingredient");
+// make this ^ live in scripts.
+
+//VIEWING SAVED RECIPES:
+
+//SAVING A RECIPE
 
 //ON PAGE LOAD
-document.addEventListener("DOMContentLoaded", (event) => {
-  displayRecipes(recipeData);
-  displayTags(recipeData);
-});
 
-function displayRecipes(array) {
+const recipeDisplay = document.querySelector(".recipes");
+const tagButtons = document.querySelector(".tag-buttons");
+
+export function displayRecipes(array, innerText) {
   let recipeHTML = "";
   array.forEach((recipeEl) => {
-    recipeHTML += `<div class="recipe-card">
-    <div class="title-recipe" id=${recipeEl.id}>${recipeEl.name}</div>
+    recipeHTML += `<div class="recipe-card"><div class="title-recipe" id=${recipeEl.id}>${recipeEl.name}</div>
     <img
       src="${recipeEl.image}"
       alt="recipe-img"
       id=${recipeEl.id}
     />
-    <button class="save-recipe-btn" id${recipeEl.id}>Save Recipe</button>
+    <button class="save-recipe-btn">${innerText}</button>
    </div>`;
   });
   recipeDisplay.innerHTML = recipeHTML;
 }
 
-function displayTags(array) {
+export function displayTags(array) {
   const tagsArray = returnListOfUniqueTags(array);
   let tagsHtml = "";
   tagsArray.forEach((tagEl) => {
@@ -83,117 +77,10 @@ function displayTags(array) {
   tagButtons.innerHTML = tagsHtml;
 }
 
-//SEARCHING FOR A RECIPE BY NAME
-// input.addEventListener("keydown", (event) => {
-//   if (event.key === "Enter") {
-//     const inputValue = input.value;
-//     const recipeSearch = returnFilteredListName(recipeData, inputValue);
-//     displayRecipes(recipeSearch);
-//   }
-// });
+// this can stay in here
 
-inputName.addEventListener("keydown", (event) => {
-  if (event.key === "Enter") {
-    const userInput = getUserInput(".input-name");
-    const recipeIdsByName = findRecipeByName(userInput, recipeData);
-    displayRecipes(recipeIdsByName);
-  }
-});
+// USING THE SEARCH BAR:
 
-inputIngredient.addEventListener("keydown", (event) => {
-  if (event.key === "Enter") {
-    const userInput = getUserInput(".input-ingredient");
-    console.log(userInput);
-    const recipeIdsByIngredient = findRecipeByIngredient(
-      userInput,
-      ingredientsData,
-      recipeData
-    );
-    displayRecipes(recipeIdsByIngredient);
-  }
-});
-
-// const displayRecipeByIds = (recipeIds) => {
-//   const recipesToDisplay = recipeData.filter((recipe) => {
-//     return recipeIds.includes(recipe.id);
-//   });
-//   let recipeHTML = "";
-
-//   recipesToDisplay.forEach((recipeEl) => {
-//     recipeHTML += `<div class="recipe-card">
-//       <img
-//         src="${recipeEl.image}"
-//         alt="recipe-img"
-//         id=${recipeEl.id}
-//       />
-//       <button class="save-recipe-btn">Save Recipe</button>
-//     </div>`;
-//   });
-//
-//   recipeDisplay.innerHTML = recipeHTML;
-// }
 //CLICKING A TAG ELEMENT
-tagButtons.addEventListener("click", (event) => {
-  let tagClicked;
-  tagClicked = event.target.id;
-  const filteredRecipeIDByTag = returnFilteredTag(recipeData, tagClicked);
-  const filteredArrayByTagID = returnFilteredRecipeArrayByTagID(
-    filteredRecipeIDByTag,
-    recipeData
-  );
-  displayRecipes(filteredArrayByTagID);
-});
 
 // CLICKING A RECIPE OR A RECIPE NAME
-
-recipeDisplay.addEventListener("click", (event) => {
-  let idClicked;
-  idClicked = event.target.id;
-  if (idClicked.length === 6) {
-    const directions = returnRecipeDirections(recipeData, idClicked);
-    const cost = returnRecipeCost(recipeData, ingredientsData, idClicked);
-    modalCost.innerText = `Estimated Cost of Ingredients: $${cost}`;
-    const ingredients = returnIngredientNames(
-      recipeData,
-      ingredientsData,
-      idClicked
-    );
-    const title = returnRecipeTitle(recipeData, idClicked);
-    modalTitle.innerHTML = title;
-
-    const tags = returnRecipeTags(recipeData, idClicked);
-    const url = returnRecipeImgUrl(recipeData, idClicked);
-
-    let directionsHtml = "";
-    directions.forEach((directionsEl, index) => {
-      let stepNumber = index + 1;
-      directionsHtml += `<li>Step ${stepNumber}: ${directionsEl}</li>`;
-    });
-    modalDirections.innerHTML = directionsHtml;
-
-    let ingredientsHtml = "";
-    ingredients.forEach((ingredientEl) => {
-      ingredientsHtml += `<li>${ingredientEl}</li>`;
-    });
-    modalIngredients.innerHTML = ingredientsHtml;
-
-    let tagsHtml = "";
-    tags.forEach((tagsEl) => {
-      tagsHtml += `<li>${tagsEl}</li>`;
-    });
-    modalTags.innerHTML = tagsHtml;
-
-    modalOverlay.classList.add("open-modal");
-
-    modalContainer.style.backgroundImage = `linear-gradient(
-      rgba(15, 15, 15, 0.7),
-      rgba(15, 15, 15, 0.7)
-    ), url(${url})`;
-  }
-});
-
-closeBtn.addEventListener("click", function () {
-  modalOverlay.classList.remove("open-modal");
-});
-
-export { displayRecipes };
